Change subject_id to string for non-integer primary keys
I have a couple of models where the primary key is not an ID, but a string (my settings table, to be exact). Trying to use the LogsActivity trait on this model causes an SQL error, as the primary key is not castable to an integer.

I don't know whether this has other consequences, but simply changing the database field type works in my application.