Make Activity model use table name from the config
The `__construct()` method is needed because PHP won’t allow to call the `config()` function when defining the `$table` property on the model.

We keep the initialization statement of the `$table` property to ensure that it will have a protected visibility as intended. Without that statement, the constructor would create this property as a public one.